Add an index to SkGlyphPos
In the mask case, the original position is needed for the fallback case.
Add an index from the original sequence to lookup the source position
when needed.

+++ src/core/SkRemoteGlyphCacheImpl.h
@@ -43,7 +43,7 @@ public:
 
     const SkGlyph& getGlyphMetrics(SkGlyphID glyphID, SkPoint position) override;
 
-    int glyphMetrics(const SkGlyphID[], const SkPoint[], int n, SkGlyphPos result[]) override;
+    size_t glyphMetrics(const SkGlyphID[], const SkPoint[], size_t n, SkGlyphPos result[]) override;
 
     bool decideCouldDrawFromPath(const SkGlyph& glyph) override;
